refactor(models): change latitude and longitude types from uint to float64
The latitude and longitude fields in the MediaByTrip struct and related service logic have been updated from uint to float64 to better represent geographic coordinates with higher precision. This change ensures accuracy in location-based operations.

+++ internal/service/media_service.go
@@ -49,8 +49,8 @@ func (s *MediaService) GetMediaByTripID(tripID int64, userID int64) ([]models.Me
 		response = append(response, models.MediaByTrip{
 			MediaID:   media.MediaID,
 			URL:       url,
-			Longitude: uint(media.GpsLongitude),
-			Latitude:  uint(media.GpsLatitude),
+			Longitude: media.GpsLongitude,
+			Latitude:  media.GpsLatitude,
 		})
 	}
 
@@ -381,7 +381,7 @@ func (s *MediaService) GetLocationByMediaID(mediaID int64) (*models.Location, er
 }
 
 func (s *MediaService) GetMediaByID(mediaID int64) (*models.Media, error) {
-    return s.MediaRepo.GetMediaByID(mediaID)
+	return s.MediaRepo.GetMediaByID(mediaID)
 }
 
 func (s *MediaService) DeleteMediaCompletely(mediaID int64, userID int64) error {
